+++ vnext/ReactUWP/Views/ControlViewManager.cpp
@@ -65,7 +65,7 @@ void ControlViewManager::UpdateProperties(ShadowNodeBase* nodeToUpdate, folly::d
       {
         if (propertyValue.isNumber())
         {
-          auto tabIndex = propertyValue.asInt();
+          auto tabIndex = propertyValue.asDouble();
           if (tabIndex == static_cast<int32_t>(tabIndex))
             control.TabIndex(static_cast<int32_t>(tabIndex));
         }
